refactor(utils): processInput 기능별로 분리
utils/processSelector.js
- 구분자 지정문을 다루는 기능들로 분리
utils/processNumbers.js
- 숫자 배열을 다루는 기능들로 분리
- getSum.js의 내용 병합
utils/checkError.js
- 에러처리를 다루는 기능으로 분리
